requirements.txt
google-generativeai
python-dotenv
PyPDF2
openai-agents